% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/betapoisson.R
\name{dbetapoisson}
\alias{dbetapoisson}
\title{Probability function of Beta-Poisson mixture distribution}
\usage{
dbetapoisson(x, a, b)
}
\description{
Probability function of Beta-Poisson mixture distribution
}
\references{
See pages 43 ff. in Jan Grandell, Mixed Poisson
    Processes, CRC Press, 1997
}

