% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/betapoisson.R
\name{kummers_m}
\alias{kummers_m}
\title{Confluent hypergeometric function of the 1st kind}
\usage{
kummers_m(a, b, z)
}
\arguments{
\item{a, b}{complex parameters in Kummer's equation}

\item{z}{complex number at which to evaluate Kummer's M function}
}
\value{
Returns the natural logarithm of the real part of Kummer's
    M function with parameters \eqn{a} and \eqn{b} evaluated at
    \eqn{z}.
}
\description{
The following differential equation is called Kummer's equation:
}
\details{
\deqn{z f''(z) + (a-z) f'(z) - b f(z) = 0}{z f"(z) + (a-z) f'(z) - b f(z) = 0}.

One of its two solutions is the confluent hypergeometric function
of the first kind, also known as Kummer's M function.
}

